# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GigGatek API
  description: |
    API documentation for the GigGatek e-commerce platform.
    
    This API provides endpoints for managing products, orders, rentals, and payments.
  version: 1.0.0
  contact:
    name: GigGatek API Support
    email: api-support@giggatek.com
    url: https://giggatek.com/api-support
  license:
    name: Proprietary
    url: https://giggatek.com/terms
servers:
  - url: https://api.giggatek.com/v1
    description: Production server
  - url: https://staging-api.giggatek.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get all orders
      description: Get a list of orders for the authenticated user. Admins can see all orders.
      operationId: getOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by order status
          schema:
            type: string
            enum: [pending, processing, shipped, delivered, cancelled, returned, refunded]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      tags:
        - Orders
      summary: Create order
      description: Create a new order
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order created successfully
                  order:
                    $ref: '#/components/schemas/OrderSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  
  # Rental Endpoints
...
  /payments/stripe/intent:
    post:
      tags:
        - Payments
      summary: Create Stripe payment intent
      description: Create a Stripe payment intent for charging a customer
      operationId: createStripePaymentIntent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  format: float
                  example: 129.97
                description:
                  type: string
                  example: GigGatek Order #123
                metadata:
                  type: object
                  properties:
                    order_id:
                      type: string
                      example: "123"
                    customer_name:
                      type: string
                      example: John Doe
                    customer_email:
                      type: string
                      example: john@example.com
                    items_count:
                      type: integer
                      example: 2
      responses:
        '200':
          description: Payment intent created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  client_secret:
                    type: string
                    example: pi_3NJkCpKXnVQrqQAZ1geFWXcR_secret_vPUjMsZRhJwRmkRPMPvYmJGtx
                  payment_intent_id:
                    type: string
                    example: pi_3NJkCpKXnVQrqQAZ1geFWXcR
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderSummary:
      type: object
      properties:
        id:
          type: integer
          example: 123
        user_id:
          type: integer
          example: 456
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled, returned, refunded]
          example: processing
        payment_status:
          type: string
          enum: [pending, paid, failed, refunded, partially_refunded]
          example: paid
        total:
          type: number
          format: float
          example: 129.97
        item_count:
          type: integer
          example: 2
        created_at:
          type: string
          format: date-time
          example: 2025-04-10T14:30:00Z
    OrderInput:
      type: object
      required:
        - items
        - shipping_address_id
        - payment_method
        - shipping_method
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - product_id
              - quantity
              - price
            properties:
              product_id:
                type: integer
                example: 101
              quantity:
                type: integer
                minimum: 1
                example: 1
              price:
                type: number
                format: float
                example: 99.99
        shipping_address_id:
          type: integer
          example: 789
        payment_method:
          type: string
          enum: [stripe, paypal]
          example: stripe
        shipping_method:
          type: string
          enum: [standard, express, overnight]
          example: standard
        notes:
          type: string
          example: Please leave package at the door
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 150
        per_page:
          type: integer
          example: 20
        current_page:
          type: integer
          example: 1
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...